Report nicely parsed errors when shader loading fails
Rather than failing as soon as possible, we continue to accumulate
errors as far into the process as we can go so the user gets the maximum
amount of information.

The error parsing needs work. I can't find a spec on WebGL specifically
but the khronos site says these error logs are implementation defined
for OpenGL, so it's likely the same on WebGL. If implementations vary
the parsing will need a lot of work.