Fix auth context failing if no user is logged in
I meant to make this defer rendering the app up we've received a
response from the backend, not wait until there is a user!

Deferring rendering will introduce some delay, but it should only happen
on the initial load. The advantage of deferring is that components know
whether a user is logged in or not on first render and so are free to
naviagte elsewhere in response without checking an additional loading
state. If the deploy proves too slow to hide this simplification, we can
revisit.