Rebuild when css.d.ts have changed
We need to rebuild here because the declaration files are generated
during a build, which means they might be out of date for tsc during
a build that produces changes to them.

This change means we sometimes build twice in a row but that is better
than having to manually force rebuilds. (At least it is right now while
build times are low).