Improve example program lighting
I was trying to decide between having a really barebones starter program
or one that produces a nice result. In the end I think it's nice if the
exmpale program has a lot of constants so people who are here to learn
can start off by just tweaking what's there.